requirements.txt
llama-index
pandas
pypdf
python-dotenv